fixed headers on philo, philo bonus updated, falta exit when times eated

+++ philo_bonus/utils.h
@@ -3,14 +3,14 @@
 # include <sys/time.h>
 # include <stdlib.h>
 
-int			ft_is_digit(char c);
-uint64_t	ft_now(void);
-int			ft_atoi(const char *str);
-int			ft_isint(const char *str);
-int			ft_strlen(const char *str);
+int				ft_is_digit(char c);
+long long int	ft_now(void);
+int				ft_atoi(const char *str);
+int				ft_isint(const char *str);
+int				ft_strlen(const char *str);
 
-int			ft_str_is_int(const char *str);
-void		ft_putstr(const char *str);
-void		ft_putnbr(uint64_t n);
+int				ft_str_is_int(const char *str);
+void			ft_putstr(const char *str);
+void			ft_putnbr(uint64_t n);
 
 #endif
